GBC functionalities: double-speed + HDMA

+++ src/PPU/PPU.h
@@ -49,10 +49,10 @@ class PPU : public Bus_obj {
   uint8_t  _OAM_SCAN_buffer[OAM_BUFFER_SIZE_BYTE];
 
   uint8_t  _DRAWING_to_wait;
-  uint16_t _DRAWING_display[SCREEN_WIDTH * SCREEN_HEIGHT];
+  uint32_t _DRAWING_display[SCREEN_WIDTH * SCREEN_HEIGHT];
   uint8_t  _DRAWING_window_condition;
   uint8_t  _DRAWING_window_line_counter;
-  uint16_t _DRAWING_display_matrix[SCREEN_HEIGHT * SCREEN_WIDTH];
+  uint32_t _DRAWING_display_matrix[SCREEN_HEIGHT * SCREEN_WIDTH];
 
   uint16_t _HBLANK_padding_to_wait;
   uint16_t _VBLANK_padding_to_wait;
@@ -69,7 +69,7 @@ class PPU : public Bus_obj {
   void reset();
   void STAT_handler(Bus_obj*);
   uint8_t get_sprite_height();
-  uint16_t get_color_from_palette(uint8_t, uint8_t);
+  uint32_t get_color_from_palette(uint8_t, uint8_t);
 
 public:
